+++ crates/micros/examples/board/avr/arduino_nano/src/bin/adc_noise.rs
@@ -16,8 +16,8 @@
 #![no_std]
 #![no_main]
 
-use devela::{BoardArduinoNano as Board, McuAtmega328p as Mcu, Radix, RandTry};
-use devela_micros::devela;
+use devela::{Radix, RandTry};
+use devela_micros::{BoardArduinoNano as Board, McuAtmega328p as Mcu, devela};
 
 devela::set_panic_handler! { loop }