+++ crates/micros/examples/board/avr/arduino_nano/src/bin/timer1_capture.rs
@@ -5,8 +5,8 @@
 #![no_std]
 #![no_main]
 
-use devela::{AvrPin, BoardArduinoNano as Board, McuAtmega328p as Mcu, is};
-use devela_micros::devela;
+use devela::is;
+use devela_micros::{AvrPin, BoardArduinoNano as Board, McuAtmega328p as Mcu, devela};
 
 devela::set_panic_handler! { loop }