+++ crates/micros/examples/board/avr/arduino_nano/src/bin/timer1_ctc.rs
@@ -5,8 +5,7 @@
 #![no_std]
 #![no_main]
 
-use devela::{BoardArduinoNano as Board, McuAtmega328p as Mcu};
-use devela_micros::devela;
+use devela_micros::{BoardArduinoNano as Board, McuAtmega328p as Mcu, devela};
 
 devela::set_panic_handler! { loop }