+++ crates/micros/src/mcu/avr/_.rs
@@ -60,6 +60,8 @@
 //
 
 crate::mods_in! {
+        #[cfg(feature = "atmega2560")]
+        mod_ atmega2560;
         #[cfg(feature = "atmega328p")]
         mod_ atmega328p;
 
@@ -86,6 +88,8 @@ crate::mods_out! { // _mods, _pub_mods, _reexports
             register::AvrReg8,
             usart::AvrUsart,
         };
+        #[cfg(feature = "atmega2560")]
+        pub use super::atmega2560::_all::*;
         #[cfg(feature = "atmega328p")]
         pub use super::atmega328p::_all::*;
     }