+++ crates/micros/src/mcu/avr/atmega2560/_.rs
@@ -1,0 +1,12 @@
+//
+//!
+//
+
+crate::mods_in! {
+    mod mcu;
+}
+crate::mods_out! { // _mods
+    _mods {
+        pub use super::mcu::McuAtmega2560;
+    }
+}